yb_graph 완성

+++ app/src/main/java/org/techtown/project/smp/smp.java
@@ -1,76 +1,76 @@
 package org.techtown.project.smp;
 
 public class smp {
-    private double future2;
-    private double future1;
-    private double present;
-    private double past1;
-    private double past2;
-    private double past3;
-    private double past6;
-    private double past12;
+    private float future2;
+    private float future1;
+    private float present;
+    private float past1;
+    private float past2;
+    private float past3;
+    private float past6;
+    private float past12;
 
-    public double getFuture2() {
+    public float getFuture2() {
         return future2;
     }
 
-    public void setFuture2(double future2) {
+    public void setFuture2(float future2) {
         this.future2 = future2;
     }
 
-    public double getFuture1() {
+    public float getFuture1() {
         return future1;
     }
 
-    public void setFuture1(double future1) {
+    public void setFuture1(float future1) {
         this.future1 = future1;
     }
 
-    public double getPresent() {
+    public float getPresent() {
         return present;
     }
 
-    public void setPresent(double present) {
+    public void setPresent(float present) {
         this.present = present;
     }
 
-    public double getPast1() {
+    public float getPast1() {
         return past1;
     }
 
-    public void setPast1(double past1) {
+    public void setPast1(float past1) {
         this.past1 = past1;
     }
 
-    public double getPast2() {
+    public float getPast2() {
         return past2;
     }
 
-    public void setPast2(double past2) {
+    public void setPast2(float past2) {
         this.past2 = past2;
     }
 
-    public double getPast3() {
+    public float getPast3() {
         return past3;
     }
 
-    public void setPast3(double past3) {
+    public void setPast3(float past3) {
         this.past3 = past3;
     }
 
-    public double getPast6() {
+    public float getPast6() {
         return past6;
     }
 
-    public void setPast6(double past6) {
+    public void setPast6(float past6) {
         this.past6 = past6;
     }
 
-    public double getPast12() {
+    public float getPast12() {
         return past12;
     }
 
-    public void setPast12(double past12) {
+    public void setPast12(float past12) {
         this.past12 = past12;
     }
 }
